SHSH is up and running with my latest ipswapi.py :P

+++ src/tss.py
@@ -4,8 +4,8 @@ import shutil
 from urllib.request import Request, urlopen
 
 try:
-    from api import API
-    from manifest import TSSManifest
+    from .ipswapi import API
+    from .manifest import TSSManifest
 except ImportError:
     raise
 
@@ -14,13 +14,14 @@ tss_headers = {'User-Agent': 'InetURL/1.0', 'Proxy-Connection': 'Keep-Alive', 'P
                'Content-type': 'text/xml; charset="utf-8"'}  # See https://www.theiphonewiki.com/wiki/SHSH_Protocol#Communication
 
 
+# TODO Add support for OTA saving
+
 class TSS(object):
-    def __init__(self, device: str, ecid: str, version='', buildid='', apnonce='', sepnonce='', bbsnum='', shsh_path='shsh'):
+    def __init__(self, device: str, ecid: str, apnonce='', sepnonce='', bbsnum='', shsh_path='shsh'):
         super().__init__()
+
         self.device = device
         self.ecid = ecid
-        self.version = version
-        self.buildid = buildid
         self.apnonce = apnonce
         self.sepnonce = sepnonce
         self.bbsnum = bbsnum
@@ -28,9 +29,9 @@ class TSS(object):
 
     def makeTSSRequest(self, data: bytes):
         request = Request(tss_url, headers=tss_headers, data=data)
-
         response = urlopen(request, timeout=5.0)
         response_text = response.read().decode('utf-8')
+
         if 'STATUS=0' not in response_text:
             if response.headers['Content-Length'] == '0':
                 print('Server returned no response... are you blacklisted?')
@@ -41,11 +42,9 @@ class TSS(object):
             return response_text[response_text.find('<?xml'):]
 
     def saveBlobs(self):
-        api = API(self.device, self.version)
-        signed_versions = api.getSignedVersions()
-
         if os.path.exists('.shsh'):
             shutil.rmtree('.shsh')
+
         os.mkdir('.shsh')
 
         if not os.path.exists(self.shsh_path):
@@ -71,39 +70,46 @@ class TSS(object):
             tss_version_response = plistlib.load(v, fmt=plistlib.FMT_XML)
 
         tss_server_version = tss_version_response['@ServerVersion']
+
         if tss_server_version != '2.1.0':
-            print('WARNING: TSS server version {} may not be supported by this script!'.format(
-                tss_server_version))
+            print(
+                f'WARNING: TSS server version {tss_server_version} may not be supported by this script!')
 
-        for stuff in signed_versions:
-            version = stuff['iOS']
-            buildid = stuff['buildid']
+        a = API(self.device)
+        signed_versions = a.getSignedVersions()
 
-            oof = API(self.device, version)
+        # FIXME I'm sure I can do this better.
 
-            # Copy the manifest inside which will contain the added string
+        for tmp1 in signed_versions.items():
+            version = tmp1[0]
+            a.version = version
+            for tmp2 in tmp1[1]:
+                if tmp2 == 'ipsw':  # FIXME Fix if there's more than one buildid
+                    buildid = tmp1[1][tmp2]['buildid'][0]
 
-            build_manifest = '.shsh/BuildManifest_{}_{}_{}.plist'.format(
-                self.device, version, buildid)
-            tss_manifest = '.shsh/TSSManifest_{}_{}_{}.plist'.format(
-                self.device, version, buildid)
+                    # Copy the manifest inside which will contain the added string
 
-            oof.downloadFileFromArchive('BuildManifest.plist', '.shsh')
-            shutil.move('.shsh/BuildManifest.plist', build_manifest)
+                    build_manifest = f'.shsh/BuildManifest_{self.device}_{version}_{buildid}.plist'
+                    tss_manifest = f'.shsh/TSSManifest_{self.device}_{version}_{buildid}.plist'
 
-            print('Converting from BuildManifest to TSS manifest...')
-            apnonce = manobj.initFromBuildManifest(
-                self.device, tss_manifest, build_manifest, self.ecid, apnonce=self.apnonce, sepnonce=self.sepnonce, bbsnum=self.bbsnum)['apnonce']
+                    a.readFromRemoteArchive(
+                        'BuildManifest.plist', True, '.shsh')
+                    shutil.move('.shsh/BuildManifest.plist', build_manifest)
 
-            print('Sending TSS request for {} ({})...'.format(version, buildid))
+                    print('Converting from BuildManifest to TSS manifest...')
+                    apnonce = manobj.initFromBuildManifest(
+                        self.device, tss_manifest, build_manifest, self.ecid, apnonce=self.apnonce, sepnonce=self.sepnonce, bbsnum=self.bbsnum)['apnonce']
 
-            with open(tss_manifest, 'rb') as f:
-                tss_response = self.makeTSSRequest(f.read())
+                    print(f'Sending TSS request for {version} ({buildid})...')
 
-            blob_path = os.path.join(self.shsh_path, '{}_{}_{}-{}_{}.shsh2'.format(
-                self.ecid, self.device, version, buildid, apnonce))
-            with open(blob_path, 'w+') as blob:
-                blob.write(tss_response)
-                print('Saved {} blob to {}'.format(version, blob_path))
+                    with open(tss_manifest, 'rb') as f:
+                        tss_response = self.makeTSSRequest(f.read())
 
-        shutil.rmtree('.shsh')
+                    blob_path = os.path.join(
+                        self.shsh_path, f'{self.ecid}_{self.device}_{version}-{buildid}_{apnonce}.shsh2')
+
+                    with open(blob_path, 'w+') as blob:
+                        blob.write(tss_response)
+                        print('Saved {} blob to {}'.format(version, blob_path))
+
+                    shutil.rmtree('.shsh')
